Updated type of User Id in Order

+++ src/main/java/org/factoriaf5/digital_academy/funko_shop/order/OrderController.java
@@ -49,7 +49,7 @@ public class OrderController {
     }
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<List<OrderDTO>> getOrdersByUser(@PathVariable Long userId) {
+    public ResponseEntity<List<OrderDTO>> getOrdersByUser(@PathVariable Integer userId) {
         List<OrderDTO> orders = orderService.getOrdersByUser(userId);
         return new ResponseEntity<>(orders, HttpStatus.OK);
     }
